+++ src/api/bookmark.js
@@ -1,6 +1,5 @@
 import axios from "axios";
 
-// const API_URL = "http://localhost:5000/posts";
 const API_URL = "http://localhost:5001/posts";
 
 const fetchPlaces = async () => {
