+++ src/pages/join/Login.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { Link, useNavigate } from "react-router-dom";
-import AuthForm from "../components/AuthForm";
-import { getUserProfile, login } from "../api/auth";
-import useAuthStore from "../zustand/authStore";
+import AuthForm from "../../components/AuthForm";
+import { getUserProfile, login } from "../../api/auth";
+import useAuthStore from "../../zustand/authStore";
 
 const Login = () => {
   const navigate = useNavigate();
