+++ src/shared/Router.jsx
@@ -2,17 +2,17 @@ import { BrowserRouter, Route, Routes } from "react-router-dom";
 import ProtectedRoute from "../components/ProtectedRoute";
 import BookmarkTestHome from "../pages/bookmarkTest/bookmarkTestHome";
 import BookmarkTestMyPage from "../pages/bookmarkTest/bookmarkTestMyPage";
-import Layout from "../components/Layout";
-import Home from "../pages/Home";
-import Login from "../pages/Login";
-import Signup from "../pages/Signup";
-import MyPage from "../pages/MyPage";
+import MainLayout from "../components/MainLayout";
+import Home from "../pages/main/Home";
+import Login from "../pages/join/Login";
+import Signup from "../pages/join/Signup";
+import MyPage from "../pages/myPage/MyPage";
 
 const Router = () => {
   return (
     <BrowserRouter>
       <Routes>
-        <Route path="/" element={<Layout />}>
+        <Route path="/" element={<MainLayout />}>
           <Route index element={<Home />} />
           <Route path="/login" element={<Login />} />
           <Route path="/signup" element={<Signup />} />
@@ -33,6 +33,7 @@ const Router = () => {
             }
           />
         </Route>
+
         <Route
           path="/mypage"
           element={
